feePerGas uint48

+++ packages/protocol/contracts/L1/TaikoData.sol
@@ -41,7 +41,7 @@ library TaikoData {
     }
 
     struct StateVariables {
-        uint64 feePerGas;
+        uint48 feePerGas;
         uint64 genesisHeight;
         uint64 genesisTimestamp;
         uint64 numBlocks;
@@ -108,7 +108,7 @@ library TaikoData {
         bytes32 metaHash;
         uint64 blockId;
         uint64 proposedAt;
-        uint64 feePerGas;
+        uint48 feePerGas;
         uint32 gasLimit;
         address proposer;
         uint24 nextForkChoiceId;
@@ -131,7 +131,7 @@ library TaikoData {
     struct Bid {
         address prover;
         uint64 deposit;
-        uint64 feePerGas;
+        uint48 feePerGas;
         // In order to refund the diff betwen gasUsed vs. blockMaxGasLimit
         // It is also part of the bidding - how fast some can submit proofs
         // according to his/her own commitment.
@@ -173,7 +173,7 @@ library TaikoData {
         uint64 nextEthDepositToProcess;
         // Slot 9
         uint64 lastVerifiedAt;
-        uint64 feePerGas;
+        uint48 feePerGas;
         uint64 lastVerifiedBlockId;
         uint64 avgProofWindow;
         // Reserved
