+++ packages/protocol/test/TaikoL1TestBase.t.sol
@@ -30,7 +30,7 @@ abstract contract TaikoL1TestBase is Test {
     // 1 TKO --> it is to huge. It should be in 'wei' (?).
     // Because otherwise first proposal is around: 1TKO * (1_000_000+20_000)
     // required as a deposit.
-    uint64 feePerGas = 10;
+    uint48 feePerGas = 10;
     uint64 proofWindow = 60 minutes;
     uint64 l2GasExcess = 1e18;
